Delete DVCmp flag
Summary: This flag is latched to false, post-specialization. Remove it and const-prop its value. In a few places, like where the flag is actually used in array-data, we can simplify the code further because when specialization is on, all the notice blocks are eliminated. We also delete two test config lines that enable this flag - they're unnecessary.

Reviewed By: alexeyt

Differential Revision: D22110928

fbshipit-source-id: 1d64bbb22c565242f5aca58868f6cf0796b0f3ce