Delete varray promotion logic
Summary:
Not only can we now delete the two varray promotion flags, we can also clean up and consolidate get/set logic that's now identical for varrays and vecs. (Things are a lot simpler because of our nice invariant: PackedKind => varray!)

We're this close to unifying LvalInt and LvalStr for varrays and vecs, too, but they have slightly different error messages. I'm going to change that in a follow-up diff - I didn't want any semantics changes in this one. (We could switch on type to change the error message, but I want to re-use MutableOpInt for that purpose and it needs to throw OOB for its other callers.)

Reviewed By: ricklavoie

Differential Revision: D22117674

fbshipit-source-id: 77fd88d5848e6a4781b67f67ee097dac37e49136