Disallow overriding of non-abstract constants defined in interfaces
Summary:
HHVM does not support one specific case of overriding constants:
If the original constant was defined as non-abstract in an interface, it cannot be overridden when implementing or extending that interface.

I replaced the existing logic for checking constant overrides, which just checked that we do not override a constant defined in multiple parent-interfaces.

Reviewed By: andrewjkennedy

Differential Revision: D22018530

fbshipit-source-id: f6c13513a194997e0e72cf72fca2c3d233b57baf