Add pagelet time in queue statistics
Summary: Measure how much time is spent in the pagelet queue. The main scenario we want to understand better is when a pagelet is queued without active workers. The suspicion is that HHVM instances with traffic patterns that result in higher pagelet usage introduce additional downstream delays. We have data via the `pagelet_queued` counter that this is happening so this diff identifies the extent to which it's a problem for a given host or cluster.

Reviewed By: binliu19

Differential Revision: D22131200

fbshipit-source-id: 4d0dbfaf1bc2d285e641462ce0f8fa2bcb31270f