Fix KindOfArray wonkiness in extensions
Summary:
Every instance of KindOfArray in extensions is a problem for KindOfDVArray (and for HAM in general). Fix the ones that aren't cases in a switch.

In PageServer and XboxServer, all we need to do is initialize the return array (at the point we know it's not going to be null) and update the typehint.

In Thrift spec validation, we need to use the right DataType to check against the spec. This DataType can change a lot; right now, it's KindOfArray, but soon, it'll be KindOfVArray or KindOfDArray, and post HADVAs, it'll be KindOfVec or KindOfDict. Luckily, we can encapsulate all three behaviors by calling toDataType on a live value.

memcache has the most bizarre example. They construct an invalid KindOfArray result when they need to deserialize on the main thread. I kept this wacky sentinel around but switched it to a KindOfResource. It'll explode noisily if there's something incorrect.

Reviewed By: ricklavoie

Differential Revision: D22188426

fbshipit-source-id: aeeb5922616088f181d365e34eb0770e95330ca1