Reduce the use of "$GLOBALS" in some tests.
Summary:
Specifically, for some tests which use "$GLOBALS[X]",
including with isset and unset, but NOT with "?? D",
change the uses to "\HH\global_get/set/isset/unset".

Note that "zend/good/tests/lang/execution_order.php"
shows that the order of evaluation for "$GLOBALS['a']"
and "\HH\global_get('a')" is subtly different, so I "fixed"
it, and the corresponding "expect" file.

There are still about 100 unit tests using "$GLOBALS".

Reviewed By: WizKid

Differential Revision: D22194111

fbshipit-source-id: 341c3833b4dfdbb319dc572f356a4d1f2768291e