dynProps are now a darray
Summary:
dynProps are sometimes a darray and sometimes a plain PHP array. This wonkiness can cause a small host of other potential bugs - for example, it means that for normal (non-collections) classes, ObjectData::toArray almost always returns a darray, but occasionally returns a plain array.

Let's ensure that dynProps is always a darray. The surface area of the change here is quite small. The only ways to set it are via setDynProps (I changed both callers), ObjectData::FromArray (I renamed to FromDArray and audited all callers), and the normal way, setting a new dynamic property (I changed the allocation call).

Reviewed By: jjgriego

Differential Revision: D22191990

fbshipit-source-id: c8da731f85882d60207a0dd1b22b5b2029a4c986