Stricter type-test requirements
Summary:
Currently, we'll JIT nonsense code for a type tests of unions, like CheckType<TVec|TArr>, or CheckType<TCountedVec|TStaticVec>, or...

I'm going to generate a small number of new such types as part of KindOfDVArray, so let's make the checks here stricter.

There's a small issue with the logic here for type == TStaticStr. We require a KindOfPersistentStr type byte for this type, which means we'll reject some static strings at runtime. This bug doesn't cause issues because we only use CheckType<StaticStr> for optimizations, but it means we're not getting the full effect of those optimizations. I did a quick scan of the prod TC. This bugfix won't have much impact, because most cases where we CheckType<StaticStr>, we already have a Str.

Reviewed By: ricklavoie

Differential Revision: D22209305

fbshipit-source-id: b15f04167f803143f62e74c38d5b64034ee6a91c