Delete GlobalsArray
Summary:
Goodbye, GlobalsArray. You will not be missed.

A couple notes on implementation. First off, we leave the Globals and RecordArray kinds around. Removing the kinds requires a few patches to how we do array type tests, which I'll make in a follow-up.

Second, the vsdebug extension tests (correctly) identified that we dropped $GLOBALS, since they use VarEnv::getDefinedVariables. That means that in the debugger UI, we won't show $GLOBALS anymore. I believe we still show all the globals individually, since getDefinedVariables iterates over the same NameValueTable that backed GlobalsArray. If we don't want to change the debugger UI, then an alternative is to synthetically create this array in getDefinedVariables, but that feels dishonest to me and may confuse people more than dropping it.

Reviewed By: jano

Differential Revision: D22166809

fbshipit-source-id: 0502c614cebf0dfd73c971de292d3a11f83986a8