Delete isVArrayType / isDArrayType
Summary:
I've been able to make KindOfDVArray work without these tests. Each place we use them is another annoying site to audit for HAM.

Re: migrating isArrayType to only match plain arrays: that's not going to happen. Instead, we're quite close to burning the plain array type out of existence. isArrayType is going to migrate to either isArrayLikeType or "dvarray, vec, or dict" - both of which are HAM-safe, and both of which are usable in all the callsites we're removing isVArrayType / isDArrayType from.

Reviewed By: dneiter

Differential Revision: D22209860

fbshipit-source-id: 0a6e3262bb0c8c9e5b84100e4896432b0e8891b5