Add utility function to temporarily drop context
Summary:
Same motivation as the last diff (D22050978), we need to provide a way to allow singleton pattern (done through usage of `__Memoize`) at the presence of contexts.
This diff adds two helper functions (for async and regular functions) to call a function without any context.
Needlessly to say, this is an escape hatch and it should only be used at these specific places where context does not matter. Ideally, it should always call functions that are annotated with `__NoContext`. Making this a runtime enforcement might be useful but I'm not sure if it is necessary.

Reviewed By: paulbiss

Differential Revision: D22121970

fbshipit-source-id: f4f0c7fbc30d3409ca9c0f12283c0693725afbcf