`calculate_errors` mode
Summary:
`hh_fanout calculate-errors` calculates the set of errors in the codebase.

Note that previously, we were doing something misleading: we were typechecking each changed file and adding their dependencies to the dependency graph. This can give correct results more often, but it's insufficient in general, since you need to perform a typecheck of the fanout files to get the real dependency graph delta.

Typechecking the changed files by themselves is still useful for the "prechecked files optimization" (an optimization where you color dependency edges based on whether the user directly changed the file that created them). However, we're not implementing that right now, and it adds some complexity, so this commit removes the old typecheck-changed-files system. We now only store dependency graph deltas whenever we perform a typecheck.

Reviewed By: ljw1004

Differential Revision: D21822109

fbshipit-source-id: a6303b1c2f9036b316af134f082364cece0ce0a8