Support explicit `--client-id` flag
Summary:
Currently, we use `--from` to identify the caller, and make sure that different callers get different cursors. However, the end goal is to make it so that some callers can share cursors. Those callers may be logically different according to `--from` (e.g. the IDE service and `hh` itself), but they will want to share the same `--client-id`.

This commit changes it so that an explicit `--client-id` flag can be passed in. For ease of interactive use, it defaults to the value for `--from` if not passed in explicitly.

This commit doesn't actually resolve all of the questions around sharing cursors. You would still need to share a cursor ID between two processes. How would that be done? One option is to save it to disk to communicate between processes. Another is to endow `hh_fanout` with more responsibilities: you could request a cursor corresponding to the latest state on disk, and it would query Watchman and bring itself up-to-date. This is deferred for later.

Reviewed By: ljw1004

Differential Revision: D21887258

fbshipit-source-id: 9089989ad8df02a4894d9a1338d46cfdc816a118