support HH\array_mark_legacy in constant initializers
Summary:
This version of validating constant expressions happens in a parser before namespace elaboration (there is 2nd copy of constant expresion validation logic in typechecker naming code, which is post-elaboration but not shared with emitter), and thus it depends on string comparisons to approximate namespacing.

Sadly, doing this validation before namespace elaboration doesn't seem to be quick-fixable in clean way.

For now add an extra case to accept HH\array_mark_legacy in addition to \HH\array_mark_legacy - the former is preferred form in WWW due to most code being in root namespace.

Differential Revision: D22266529

fbshipit-source-id: 9b1784621dbf5e0a0d1a1db9f1b215d3aaa7bdd3