Fix Shapes::removeKey quadratic blowup
Summary:
The following code leads to the creation of a type that is O(n^2) in size.
```
type Tshape = shape(?'f1' => int, ..., ?'fn' => int);
function test():bool { ... }
function bad():Tshape {
  $x = shape('f1' => 42, ..., 'fn' => 42);
  if (test()) Shapes::removeKey(inout $x, 'f1');
  ...
  if (test()) Shapes::removeKey(inout $x, 'fn');
  // At this point the size of the type of `$x` is O(n^2). It has some type `shape('f1' => u1, ..., 'fn' => un)
  // where each of the types u1 to un is the union of n type variables
  return $x;
}
```
The root cause is a "freshening" operation that occurs when the special typing of `removeKey` forces the type of its parameter to be a `shape`. The types of the fields are replaced by type variables, in a subtype relationship with the original type. This is useful in general when "force solving" type variables, to prevent unnecessary freezing of types in covariant/contravariant positions (e.g. solving `vec<C> <: #1` to `#1 := vec<C>` might be too strict, if later it turns out that we want `#1 := vec<B>` for `C <: B`). But in this special case, we know that the type of the fields of the shape are not flexible: they cannot vary across the `removeKey` call.

The fix is simply to avoid the "freshening" of types in one use of `Typing_solver.expand_type_and_solve`.

Reviewed By: mpu

Differential Revision: D22254937

fbshipit-source-id: 7b0e12e95eacf17ce4c1306bb1a8a21d1f1315ca