Update 'yield from' tests
Summary:
In the initial version of this diff, I modified "yield from" tests to implement the same logic with alternative syntax instead. Doing so revealed some behavior issues, which I'll document below. In this final version, we opted to just delete "yield from" tests instead.

In most cases, we can replace "yield from" with "foreach...yield", which is clean enough. However, this diff reveals some unusual behaviors of the "yield from" syntax. For most of these, we can simulate the behavior without the syntax, at the cost of being more verbose (which is fine: the behavior is strange):

1. If users are using sent values, then they'll need to manually iterate and call send() for now.
1. If users are using the generator's return value, they'll need to save the generator in a local and call getReturn() at the end.
1. The key index counter for "yield from" is internal to the callee, but for "foreach...yield" it's external to the caller. That would require renumbering in basic_yield_from_proxying.php.
1. If users use yield from with a non-iterable, it will raise an error; foreach over a non-iterable will raise a warning instead. I think that's okay.

However, there is one behavior that I see no reasonable way of simulating without "yield from" syntax. If we "yield from" a value aliased elsewhere, and that other reference is used to call next() on it, it'll update the state of the delegate generator eagerly. This behavior is a bit of a mess. We have no plans to support it. It's tested in three tests, which I delete here: yield_from_advance_inner_generator.php, yield_from_advance_inner_iterator.php, and yield_from_multi_tree.php.

Reviewed By: jano

Differential Revision: D22300755

fbshipit-source-id: 1c5a4ee56a2a7470b071283086716b82ef6f1588