Adjust repo string length limit
Summary: The `SQLITE_LIMIT_LENGTH` value in sqlite affects both the size of TEXT and BLOB fields as well as the overall size of rows. Unfortunately I couldn't find a good mechanism to estimate the size of a row, so here I just subtract off the size of the integer field we know will also be in the row (https://www.sqlite.org/limits.html).

Reviewed By: jano, alexeyt

Differential Revision: D22261137

fbshipit-source-id: bbe39f9be3121f68be65bb501acad606427df0ea