Eliminate array typehints from ext_std_intrinsics.php
Summary:
This gets HHVM systemlib clean of `array` typehints, making it possible to ban `array` typehints in the parser (or rather in lowerer).

This eliminates test coverage for `array` typehints, but this isn't huge deal, considering that next diff in this stack will be an option to ban array typehints.

Reviewed By: billf

Differential Revision: D22327389

fbshipit-source-id: cc16e59c3a56fb9c40b141b614077ffb8f977dfd