Fix exception management in PU + traits
Summary:
We need to catch any Exception when dispatching on traits, to make
sure we account for the failing traits: since we can't check if a trait
supports an atom at compile time, we ask all of them, which might trigger
Exception. We know we are same because of the typing.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D22333225

fbshipit-source-id: 7e14cfd72d117d8c092ab1f617a448a285f6b486