Goodbye ArrayKind specializations
Summary:
We no longer generate ArrayKind specializations. Delete them and const-prop.

In addition to a cleanup, this diff is a small optimization on top of the previous one, because we lost specializations for AKExists and ArraySet when we added KindOfDVArray. We get the AKExists specialization back in irgen-builtin, and we get the ArraySet specialization back by adding some substitute logic to CallSpec::array. (We'll delete CallSpec::array in a follow-up.)

Reviewed By: jjgriego

Differential Revision: D22339439

fbshipit-source-id: 9b4a38e0f05a6fb63d35de53347823c089e777fe