Use KindOfDVArray for dvarrays
Summary:
Using a DataType to represent dvarray-ness is attractive for a few reasons:

1. For performance: it means we'll always guard to dvarray-ness, so typehint checks are basically free. We should get ~0.3% perf back with this change. We'll get other benefits, too - for example, we may specialize more iterators this way.

2. For code cleanliness: we can share a lot more JIT code between (darray and dict) and (varray and vec) and clean up specialized-type handling if we get rid of the ArrayKind specialization. Further cleanups will also be possible if we get rid of "plain arrays" (which are happening soon).

In this diff, we do close to the minimal set of changes needed to get these DataTypes working. It's still a lot of changes. Many places in the runtime need to be modified to use ArrayData::toDataType() or to produce different DataTypes; see D22336137 for just that part. In the JIT, we make TArr a union type of TPArr | TVArr | TDArr, and adjust a lot of code; see D22336158 for just that part. We need to land the two changes together because landing only the runtime changes breaks assumptions all over the JIT.

There's more we can do afterwards, but (in order to keep this review from getting even bigger) I've split it out. Some cleanups are: 1) get rid of CheckVArray et. al., 2) get rid of ArraySpec::kind() completely (it's already dead as of this change) and 3) update the vec and dict IR ops to accept varray and darray, respectively, and improve codegen a bunch more.

Reviewed By: jjgriego

Differential Revision: D22196049

fbshipit-source-id: d564315d0bd8cc262d8f8efa6e1a1d654885dc7f