Ban new classname with reified generics
Summary:
Instantiating a class from a `classname<class>` in which the class has reified generics causes a runtime error. However, it currently does not throw a type error. We do not want to error on `ReifiedClass::class` because it is still usable as a string and for static methods.

We want to error on the `new` portion of it.

Reviewed By: Wilfred

Differential Revision: D22145152

fbshipit-source-id: de1c684d023cf9da5bf1cba3042dac8c70d7ebd4