Don't require explicit `state#save` calls
Summary: Currently, every call to `Incremental.state#add_cursor` needs to be followed up with a call to `Incremental.state#save` to actually persist it to disk. It might cause a problem in the future if someone forgets to call `#save`, or — more likely — `#save` fails to be called due to some kind of refactoring and thrown exception.

Reviewed By: ljw1004

Differential Revision: D22052760

fbshipit-source-id: 5cb82a5fdccfb8756185bf8f0829c82302f55e8b