Calculate fanout when advancing cursor
Summary:
When files are changed on disk, a new cursor is created which stores which files have changed. When we want to `calculate-errors` for the codebase, we must take all the file changes since the last `calculate-errors` operation, calculate their fanout, and then typecheck the fanout.

We could calculate the fanout at one of two times:

1) When we are about to `calculate-errors`, read the symbols from disk for every changed file and calculate fanout.
2) Whenever we advance the cursor with `calculate`, we can read the files from disk then and calculate the fanout immediately, and store it for later.

Currently, we *already* calculate the fanout on `calculate` and display it to the user, and just don't store it. Then we have to recalculate it later. We may as well save the results.

Storing the fanout in this way has pragmatic consequences: it should be much easier to debug what happened in a sequence of events given a state file. Otherwise, we won't have stored the calculated fanout in the database on-disk at all!

Note that neither approach fully protects against race conditions: if a file changes on disk during `calculate` or `calculate-errors`, the results are not reliable. The calculation should be thrown away and retried. The timing windows for race conditions are different, but still exist for both.

Reviewed By: ljw1004

Differential Revision: D22052759

fbshipit-source-id: 087b027867de8da7200c0616cc63928e4e4c692b