array_merge returns darray
Summary:
array_merge is still rather common in our code. I'm honestly not sure why we didn't do this one earlier.

The only real trick here is that we have to be "HAM-safe" - we need to return a dict post-HADVAs. We also can't change the typehint because of the null return case. (Both a ?darray and a darray typehint are valid here - HHVM expands array builtin return types to include null - but the problem is the use of the sneaky getCheckedContainer macro. I don't care to fix it.)

Reviewed By: ricklavoie

Differential Revision: D22371748

fbshipit-source-id: de67e62f9951e149240dc2809f5a3bda3fbee374