Fake member bookkeeping simplification
Summary:
Primary change that doesn't effect the behaviour: instead of keeping a ledger with two constructors, have one constructor with valid and invalid sets (same as the now deleted `Invalidated` constructor).

Behaviour change: when joining two ledgers, the invalid set is not just a union of individual invalid sets, but also the addition of elements that are outside the intersection of valid sets. They are invalidated because they go out of scope.

This is consistent with how a `Valid` and `Invalidated` ledgers were joined, but not with two `Invalidated` ledgers. This was an inconsistency and the change fixes it in favour of keeping maximal information.

The change should improve some error messages by pointing to a reason for a previously refined fake member that went out of scope.

Reviewed By: andrewjkennedy, vsiles

Differential Revision: D22333075

fbshipit-source-id: 7535dff14a56bc00e2c0814678c380579b72713a