Invalidate invalid fake members
Summary:
A fake member should not be both invalid and valid at the same time. This invariant is stated in `typing_fake_members.ml` but not enforced.

The following is enough to make a fake member both invalid and valid:
```
$c->i = 42;
foo();
$c->i = 24;
hh_show_env();
```

This diff enforces the invariant by removing invalid occurrences of a fake member from the invalid set when we add it as a valid fake member.

Reviewed By: CatherineGasnier

Differential Revision: D22333143

fbshipit-source-id: 5b3238e5c525cc227f2d7f8971f663fb013e07f9