Populate naming cache based on decl results and vice versa
Summary: If we get a decl from a decl request, we can determine what file it is in, so we can populate the naming cache and avoid an IPC round-trip. Likewise, if we perform a naming request and learn that there is *no* decl, we can save that information in the decl cache and avoid future round-trips.

Reviewed By: Wilfred

Differential Revision: D20527978

fbshipit-source-id: eff923f2560fe6197227b489412581ef8ab777ee