Speculative decl-parsing
Summary: When a naming table API is used, respond to the client, then speculatively decl-parse the file and add its decls to the cache before listening for the next request. When responding to a decl request, add the non-requested decls in the file containing the requested decl to the cache.

Reviewed By: Wilfred

Differential Revision: D20503257

fbshipit-source-id: 48fec1ca60b3dd4627a3bb1bb920b4d0612f6414