Add case-insensitive function and type lookup
Summary: To service the get-canon-name APIs (which return the "canonical" casing of a symbol name), we need a case-insensitive means of looking up symbols from the naming table database. We can't just use COLLATE NOCASE, since we look up names using a hash. Instead, this diff adds an API which takes an owned String, mutates it to be (ASCII) lowercase, then uses that lowercased symbol name to produce the hash.

Reviewed By: Wilfred

Differential Revision: D20527979

fbshipit-source-id: 62a91c67b883a24cf0cf86ea60fb1e896220d5ee