Preserve sharing for references and Rc
Summary:
While sharing is preserved for some of our data structures using the intrusive OCaml-value cache in RcOc, we'd like to avoid reference counting or intrusive caching in our arena-allocated structures, and simply use references instead. Since we do not currently preserve sharing for references, this results in filenames being copied many times over when converting a decl to OCaml.

I'd like to preserve sharing for references by reviving the approach I abandoned in {D17864958}. The allocator is expected to maintain a memoization table for conversions, and the implementations of ToOcamlRep for &T and Rc<T> ask the allocator to return the cached value if it exists (by invoking `Allocator::memoized`).

I abandoned this approach previously because I had concerns about the lifetime of the memoization table (see the comments on D17864958). If it lived as long as the Arena, then reuse of memory addresses or mutations could lead to `Allocator::memoized` returning stale results (see the added test case `mutated_reference` for an example). I've tried to address this by only memoizing conversions (and thus preserving sharing) if the caller invokes `Allocator::add_root` instead of `Allocator::add`, and requiring that the memoization table lives only for the duration of the `add_root` call. The type system guarantees that the converted value won't be mutated or dropped during the `add_root` call (because `add_root` borrows it), so we don't need to worry about memory addresses being reused or values being mutated for any of the values reachable from the conversion root.

An implementor of ToOcamlRep might observe stale results by allocating a new value, converting it via `Allocator::add`, mutating it, and then converting it again, so I've noted in the ToOcamlRep docs that implementors are expected not to mutate or drop (since the same problem can arise for a new value allocated at a dropped value's address) values that they convert. Implementations of ToOcamlRep shouldn't need to do these things anyway, I think.

Reviewed By: shiqicao

Differential Revision: D21556226

fbshipit-source-id: 7d0f90509db113fb7274eeefa9f5f1d9d4637cc2