Scoped guards for func body accesses
Summary:
"Hide" the block data of a function by renaming the field to rawBlocks. This member will contain some kind of serialized data instead of our current php::Block representation. Use scoped smart pointers to represent deserialization, so that we can expand the bytecode, perform operations, and then drop the extra memory.

This diff is almost entirely noise; interesting parts are in context.h and representation.h, and every call to a constructor of ConstFunc or MutFunc.

The reason to have two such types is to avoid conflating the const-ness of the pointer type with the const-ness of its target. That is, it's insufficient to have a ExpandedFunc type with "const Func* operator->() const" and "Func* operator->()", because we can copy a const ExpandedFunc to a mutable ExpandedFunc and then use the mutable accessor. Instead, it's safer to decide the const-ness of the smart pointer's target based on the const-ness of the raw pointer used to initialize the wrapper.

The places that we're going to do deserialization are precisely those places that we instantiate these smart pointers. It's important that we limit these. Here's the list:
- ConstFunc is instantiated in:
  - check.cpp and show.cpp, just for debugging
  - emit.cpp, when we're ready to actually write out the data
- MutFunc is instantiated in:
  - func-util.cpp and index.cpp, for copying data from one Func to another (e.g. traits)
  - stats.cpp, for debugging (?)
  - parse.cpp, when it's time to construct the Func
  - whole-program.cpp, when it's time to free the Func

Finally, there's the big one: we instantiate a MutFunc as part of the AnalysisContext used throughout HHBBC whenever we're actually dealing with one particular Func. This bit is the main purpose of the change, really: we want to keep all the Funcs compressed except the 60-odd Funcs that are currently under consideration by program threads.

I want to make a note about thread-safety and deserialization. Now that we have this abstraction barrier, we can store the raw data within Func in pretty much any format we'd like. I'll experiment with a few formats. When we do instantiate a ConstFunc or MutFunc and trigger scoped deserialization, we'll increment an atomic counter within the Func. Only if we bump from 0 to 1 will we deserialize; any other depths will share access to the mutable data. That's why we needed an extra pointer in Func (see previous diff) - the Func will point to both compact and deserialized data.

Reviewed By: mofarrell

Differential Revision: D22329012

fbshipit-source-id: dd8d60f1d12c42209ebd3e934f2db129b0b9f85e