Revert `check_kind_to_string` to return variant names
Summary:
This function was meant to be an enum/variant `to_string` function, but it was also used in logging, which was confusing.

This diff reverts the `to_string` function to return just the variant constructor names, and changes the logging to use the strings from the previous diff directly.

Reviewed By: ljw1004

Differential Revision: D22402531

fbshipit-source-id: dfb118002a678f61ec8407d888e9eba5893fa302