Harden dvarray bytecode invariants
Summary:
Reified generics are managed by the runtime and cannot be written to by Hack directly. The reified generics local is always a HAM-safe varray (i.e. post-HADVAs, a vec). Similarly, varargs are always a HAM-safe varray. Modify the JIT and interpreter to assert or check this requirement where it previously only tested for arrays.

The stricter checks look like a behavior change, but we never generate bytecode that violates them, and I don't care to write custom HHAS to trigger the error.

Reviewed By: ricklavoie

Differential Revision: D22407433

fbshipit-source-id: 75d94c4e26758837980e8598308abeb59792806a