Use hh_decl canon name APIs in Decl_service backend for Naming_provider
Summary:
Duplicate name error detection requires the ability to perform a case-insensitive lookup of the definition position of functions and types (since symbol references are case-insensitive in the runtime). Since worker processes do not have access to the naming table (they could mount the database, but we do not plan to copy a naming table delta into each worker), they must get this information from the decl service.

This diff hooks up the Decl_service provider to the APIs added in the previous diff.

Reviewed By: Wilfred

Differential Revision: D20527982

fbshipit-source-id: c76e03be54148acfc011019c500e68fe48a6239c