Rip out plain array bytecodes
Summary:
The parser can still parse "array", but we'll raise an error on compiling bytecode for it.

I'm making the parser change separately because a) there are parser tests that fail and b) it's so boilerplate-y with ~50 files to change. Perhaps codegen will be easier than it looks.

Deleted: CastArray, NewArray, NewMixedArray, NewPackedArray, NewStructArray

Still present: Array (used for static darrays and varrays). We eliminate Array from the parser/emitter's TypedValue union and verify that Array always produces a darray.

Reviewed By: ricklavoie, jano

Differential Revision: D22420376

fbshipit-source-id: f013062b63a4fb8e95493b9f2282fd44cfe479d6