De-template ArrayData::convertKey
Summary:
Rename ArrayData::convertKey to ArrayData::intishCastKey and drop the template parameter. Only call it from those sites that cast before. In a few places, we call this method with the default IntishCast::None - these callers are all trivial, so we can simplify that code.

Note that Array::convertKey does other conversions in addition to intish-cast, so we can't clean it up so easily. We'll address it in a follow-up.

Reviewed By: ricklavoie, dneiter

Differential Revision: D22414723

fbshipit-source-id: b5bfa043f5ad89569e85531dbf18cad3d3e51105