Log on nontrivial intish casts
Summary:
I'm sure that these logs are still happening. These functions force us to observe PHP arrays within builtins, so I'd like to at least eliminate as many builtin users as possible.

We don't technically need to log in the standalone case for HAM, but why not try to kill it too?

Reviewed By: ricklavoie

Differential Revision: D22407271

fbshipit-source-id: dabd30f0552ea0ee9a4dddcbd98c8d381ed98a68