correctly identify information flow due to early return in an IF statement
Summary:
If one branch of an if statement `return`, then information about the program counter leaks into any code executing after the IF

To address this, I've added the PC info to an exit continuation which gets added to the Next continuation after the if statement.

I'm not sure if this is rigorous, but it does work on my test case.

Reviewed By: mpu

Differential Revision: D22357395

fbshipit-source-id: f08e0b35c19a5fbd5755d9d44f7a67abe03fe0a4