handle uncaught exceptions differently
Summary:
I noticed from logs that if an ocaml binary encounters an unhandled exception, then it exits with exit code 2. But that's the same exit code we use from `hh_client check` to indicate that things worked fine and there were typechecker errors!

I saw this in perf logs, where we incorrectly believed that the perf run had finished quickly (with typechecker errors) whereas actually it had hit an unhandled exception.

There's no way we can change all the clients in the world to recognize a different exit code out of hack. We're stuck with "2" indicating success with hack errors.

So what I've done in this diff is (1) make sure that we handle otherwise-unhandled exceptions from hh_client, (2) we use exit code 221 for them.

Also, as documented in the code, I trust that whenever the code uses Exit_with then it has the decency to have printed some user-facing message if relevant. I only made the top-level exception handler print a message in case of uncaught exceptions. That's because of cases like `Exit_with Server_hung_up_should_retry`. I wouldn't want to display error messages and callstacks in this case, since it's a normal and expected part of behavior.

CatherineGasnier - I expect this change to reduce variability in our perf runs!

Reviewed By: 2BitSalute

Differential Revision: D22364920

fbshipit-source-id: 31d16357a01a901b8580ad3c00e1305835667199