Throw inside HHVM when indexing into an array using a non-arraykey type
Summary:
PHP arrays have for the longest time allowed you to use non-arraykey types as array keys, and silently coercing anything that's not a string or an int into one.

Notices for this behavior have been available behind the following flags:
```
-vRuntime.Eval.HackArrCompatNotices=1
-vRuntime.Eval.HackArrCompatCheckArrayKeyCast=1
```

As part of the effort to unifying hack arrays and PHP arrays, we are getting rid of this implicit coercion functionality, such that trying to index into a PHP array using a non-arraykey type will now throw. This matches the behavior for hack arrays.

Differential Revision: D19824967

fbshipit-source-id: 93ec5b4055ad003e6616e1600385e7c76da9c2bd